flake: switch to `bootspec-secureboot` fork
Switching to this fork is necessary in order for secureboot to properly
work with newer systemd versions.  In the future I plan to switch to
[lanzaboote](https://github.com/nix-community/lanzaboote), but their
stable release doesn't have the changes required for the new systemd
version, and following their master branch leads to a lot of rebuilds
which I'd like to avoid.

Signed-off-by: Matthew Penner <me@matthewp.io>